fix: various issues with discovery + testing
