+++ chord_metadata_service/discovery/censorship.py
@@ -1,6 +1,6 @@
 import sys
 
-from .config import discovery_config
+from django.conf import settings
 
 __all__ = [
     "RULES_NO_PERMISSIONS",
@@ -26,9 +26,11 @@ def get_threshold(low_counts_censored: bool) -> int:
     """
     Gets the maximum count threshold for hiding censored data (i.e., rounding to 0).
     """
-    if not discovery_config:
+    if not low_counts_censored:
+        return 0
+    if not settings.CONFIG_PUBLIC:
         return RULES_NO_PERMISSIONS["count_threshold"]
-    return discovery_config["rules"]["count_threshold"] if low_counts_censored else 0
+    return settings.CONFIG_PUBLIC["rules"]["count_threshold"]
 
 
 def thresholded_count(c: int, low_counts_censored: bool) -> int:
@@ -39,6 +41,8 @@ def get_max_query_parameters(low_counts_censored: bool) -> int:
     """
     Gets the maximum number of query parameters allowed for censored discovery.
     """
-    if not discovery_config:
+    if not low_counts_censored:
+        return sys.maxsize
+    if not settings.CONFIG_PUBLIC:
         return RULES_NO_PERMISSIONS["max_query_parameters"]
-    return discovery_config["rules"]["max_query_parameters"] if low_counts_censored else sys.maxsize
+    return settings.CONFIG_PUBLIC["rules"]["max_query_parameters"]
