+++ chord_metadata_service/restapi/tests/constants.py
@@ -366,195 +366,3 @@ INDIVIDUALS_NOT_ACCEPTED_DATA_TYPES_LIST = [
 INDIVIDUALS_NOT_ACCEPTED_DATA_TYPES_DICT = [
     {**item, "extra_properties": extra_properties_with_dict} for item in deepcopy(VALID_INDIVIDUALS)
 ]
-
-
-CONFIG_PUBLIC_TEST = {
-    "overview": [
-        {
-            "section_title": "First Section",
-            "charts": [
-                {"field": "age", "chart_type": "bar"},
-                {"field": "sex", "chart_type": "pie"},
-            ]
-        },
-        {
-            "section_title": "Second Section",
-            "charts": [
-                {"field": "date_of_consent", "chart_type": "bar"},
-                {"field": "smoking", "chart_type": "bar"},
-                {"field": "baseline_creatinine", "chart_type": "bar"},
-            ]
-        },
-        {
-            "section_title": "Third Section",
-            "charts": [
-                {"field": "lab_test_result_value", "chart_type": "bar"},
-            ]
-        }
-    ],
-    "search": [
-        {
-            "section_title": "First Section",
-            "fields": [
-                "sex", "age", "smoking", "covidstatus", "death_dc",
-                "lab_test_result_value", "baseline_creatinine", "date_of_consent",
-                "tissues"
-            ]
-        }
-    ],
-    "fields": {
-        "sex": {
-            "mapping": "individual/sex",
-            "title": "Sex",
-            "description": "Sex at birth",
-            "datatype": "string",
-            "config": {
-                "enum": None
-            }
-        },
-        "age": {
-            "mapping": "individual/age_numeric",
-            "title": "Age",
-            "description": "Age at arrival",
-            "datatype": "number",
-            "config": {
-                "bin_size": 10,
-                "taper_left": 10,
-                "taper_right": 100,
-                "units": "years",
-                "minimum": 0,
-                "maximum": 100
-            }
-        },
-        "smoking": {
-            "mapping": "individual/extra_properties/smoking",
-            "title": "Smoking",
-            "description": "Smoking exposure",
-            "datatype": "string",
-            "config": {
-                "enum": [
-                    "Non-smoker",
-                    "Smoker",
-                    "Former smoker",
-                    "Passive smoker",
-                    "Not specified"
-                ]
-            }
-        },
-        "covidstatus": {
-            "mapping": "individual/extra_properties/covidstatus",
-            "title": "Covid status",
-            "description": "Covid status",
-            "datatype": "string",
-            "config": {
-                "enum": [
-                    "Positive",
-                    "Negative",
-                    "Indeterminate"
-                ]
-            }
-        },
-        "death_dc": {
-            "mapping": "individual/extra_properties/death_dc",
-            "title": "Death",
-            "description": "Death status",
-            "datatype": "string",
-            "config": {
-                "enum": [
-                    "Alive",
-                    "Deceased"
-                ]
-            }
-        },
-        "lab_test_result_value": {
-            "mapping": "individual/extra_properties/lab_test_result_value",
-            "title": "Lab Test Result",
-            "description": "This acts as a placeholder for numeric values",
-            "datatype": "number",
-            "config": {
-                "bins": [200, 300, 500, 1000, 1500, 2000],
-                "minimum": 0,
-                "units": "mg/L"
-            }
-        },
-        "baseline_creatinine": {
-            "mapping": "individual/extra_properties/baseline_creatinine",
-            "title": "Creatinine",
-            "description": "Baseline Creatinine",
-            "datatype": "number",
-            "config": {
-                "bin_size": 50,
-                "taper_left": 50,
-                "taper_right": 200,
-                "minimum": 30,
-                "maximum": 600,
-                "units": "mg/L"
-            }
-        },
-        "date_of_consent": {
-            "mapping": "individual/extra_properties/date_of_consent",
-            "title": "Verbal consent date",
-            "description": "Date of initial verbal consent(participant, legal representative or tutor), yyyy-mm-dd",
-            "datatype": "date",
-            "config": {
-                "bin_by": "month"
-            }
-        },
-        "tissues": {
-            "mapping": "biosample/sampled_tissue/label",
-            "mapping_for_search_filter": "individual/biosamples/sampled_tissue/label",
-            "title": "Tissue",
-            "description": "Tissue from which the biosample was extracted",
-            "datatype": "string",
-            "config": {
-                "enum": None
-            }
-        }
-    },
-    "rules": {
-        "count_threshold": 5,
-        "max_query_parameters": 2
-    }
-}
-
-CONFIG_PUBLIC_TEST_SEARCH_SEX_ONLY = deepcopy(CONFIG_PUBLIC_TEST)
-CONFIG_PUBLIC_TEST_SEARCH_SEX_ONLY["search"][0]["fields"] = ["sex"]
-
-CONFIG_PUBLIC_TEST_SEARCH_UNSET_FIELDS = deepcopy(CONFIG_PUBLIC_TEST)
-CONFIG_PUBLIC_TEST_SEARCH_UNSET_FIELDS["fields"].update([
-    ("unset_date",
-     {
-         "mapping": "individual/extra_properties/unset_date",
-         "title": "Some date",
-         "description": "Some date",
-         "datatype": "date",
-         "config": {
-             "bin_by": "month"
-         }
-     }),
-    ("unset_numeric",
-     {
-         "mapping": "individual/extra_properties/unset_numeric",
-         "title": "Some measure",
-         "description": "Some measure",
-         "datatype": "number",
-         "config": {
-             "bin_size": 50,
-             "taper_left": 50,
-             "taper_right": 500,
-             "minimum": 0,
-             "maximum": 600,
-             "units": "mg/L"
-         }
-     }),
-    ("unset_category",
-     {
-         "mapping": "individual/extra_properties/unset_category",
-         "title": "Some things",
-         "description": "Some things",
-         "datatype": "string",
-         "config": {
-             "enum": None
-         }
-     })
-])
